Move GoogleLoginButton.tsx under home directory.
